+++ packages/components/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @equinor/mad-components
 
+## 0.6.9
+
+### Patch Changes
+
+-   a29275a: Change Autocomplete to use TextField instead of Input
+-   26fbbbe: added variant prop for validation handling in input component
+
+## 0.6.8
+
+### Patch Changes
+
+-   e8d96f6: Added styling fixes for Autocomplete component
+
 ## 0.6.7
 
 ### Patch Changes