+++ packages/navigation/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @equinor/mad-navigation
 
+## 0.1.14
+
+### Patch Changes
+
+-   Updated dependencies [a29275a]
+-   Updated dependencies [26fbbbe]
+    -   @equinor/mad-components@0.6.9
+
+## 0.1.13
+
+### Patch Changes
+
+-   Updated dependencies [e8d96f6]
+    -   @equinor/mad-components@0.6.8
+
 ## 0.1.12
 
 ### Patch Changes